requirements.txt
openai==1.86
python-telegram-bot==20.4
google-genai==1.20.0